prefilled forms for mult and division, fixed some styles
I fixed some styles that broke due to applying box-sizing correctly after not being applied to all items previously. I also changed all the pages to be forms so I could set each form to be submitted and generate the image on page load for the links from the home page.